# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Brazilian Artisanal Bags E-commerce API
  description: API documentation for the Brazilian Artisanal Bags E-commerce platform
  version: 0.1.0
  contact:
    email: backend@artesanato-ecommerce.com
servers:
  - url: https://artesanato-ecommerce.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
  /products:
    get:
      summary: List all products
      description: Returns a paginated list of products with optional filtering
      tags:
        - products
      parameters:
        - name: category
          in: query
          description: Filter products by category
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 12
            maximum: 50
            minimum: 1
        - name: sort
          in: query
          description: Sorting options
          required: false
          schema:
            type: string
            enum: [price_asc, price_desc, newest, oldest]
            default: newest
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of products
                      page:
                        type: integer
                        description: Current page number
                      limit:
                        type: integer
                        description: Items per page
                      pages:
                        type: integer
                        description: Total number of pages
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        inventory_count:
          type: integer
        category:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
...
    ProductImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        alt_text:
          type: string
        is_primary:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string